Set useDynamicSymbols to FALSE
It is better practice to not allow the DLL to be searched for entry
points via strings. Setting it to TRUE probably made it easier to get
things working initially, but xts still passes R CMD check when it is
set to FALSE.

Thanks to Kurt Hornik for the suggestion.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/zoo@1132 edb9625f-4e0d-4859-8d74-9fd3b1da38cb