Replace as.zoo.xts() with xts:::as.zoo.xts()
Both zoo and xts register an as.zoo.xts() method. The package that
owns the class should define the conversion method, so xts' version
should be removed.

Achim noted the difference in as.zoo.xts() between zoo and xts is
small (with xts being better), and suggested we replace the zoo
version with the code from the xts version.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/zoo@1134 edb9625f-4e0d-4859-8d74-9fd3b1da38cb